feat: refactor map components and remove unused KML data file
