+++ .env.example
@@ -1,1 +1,2 @@
 MAPTILER_API_KEY=
+KML_URL=
