+++ src/components/PreviewMap.tsx
@@ -2,6 +2,7 @@
 
 import * as mapTilerSDK from '@maptiler/sdk';
 import { type FC, useCallback, useEffect, useRef, useState } from 'react';
+import type { Road } from '@/types/road';
 import '@maptiler/sdk/dist/maptiler-sdk.css';
 import { MapStyle } from '@maptiler/sdk';
 import {
@@ -15,18 +16,14 @@ import { transformRequest } from '@/utils/mapUtils';
 
 // temporarily set api key to dummy key to remove the error
 mapTilerSDK.config.apiKey = 'abcdefghijklmnopqrstuvwxyz';
+mapTilerSDK.config.caching = true;
 
 interface MapProps {
-  center: [number, number];
-}
-
-interface Road {
-  name: string;
-  coordinates: [number, number][];
+  userLocation: [number, number] | null;
+  roads: Road[];
 }
 
 function logError(context: string, error: unknown) {
-  // Safe error logging that doesn't expose internal details
   const errorMessage = error instanceof Error ? error.message : 'Unknown error';
   console.error(`[PreviewMap] ${context}: ${errorMessage}`);
 }
@@ -39,15 +36,12 @@ function isValidLongitude(lng: number): boolean {
   return !Number.isNaN(lng) && lng >= -180 && lng <= 180;
 }
 
-const PreviewMap: FC<MapProps> = ({ center }: Readonly<MapProps>) => {
+const PreviewMap: FC<MapProps> = ({ userLocation, roads }) => {
   const mapContainer = useRef<HTMLDivElement>(null);
   const map = useRef<mapTilerSDK.Map | null>(null);
   const geolocateControl = useRef<mapTilerSDK.GeolocateControl | null>(null);
   const [mapError, setMapError] = useState<string | null>(null);
-  const [roads, setRoads] = useState<Road[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
 
-  // Helper function to initialize map
   const initializeMap = useCallback(
     (coordinates: [number, number]) => {
       if (!mapContainer.current) return;
@@ -66,20 +60,14 @@ const PreviewMap: FC<MapProps> = ({ center }: Readonly<MapProps>) => {
         transformRequest,
       });
 
-      // Initialize geolocate control
       geolocateControl.current = new mapTilerSDK.GeolocateControl({
-        positionOptions: {
-          enableHighAccuracy: true,
-        },
-        fitBoundsOptions: {
-          maxZoom: DEFAULT_ZOOM,
-        },
+        positionOptions: { enableHighAccuracy: true },
+        fitBoundsOptions: { maxZoom: DEFAULT_ZOOM },
         trackUserLocation: true,
         showAccuracyCircle: true,
         showUserLocation: true,
       });
 
-      // Add controls to map
       map.current.addControl(
         new mapTilerSDK.NavigationControl(),
         'bottom-right',
@@ -132,77 +120,41 @@ const PreviewMap: FC<MapProps> = ({ center }: Readonly<MapProps>) => {
   );
 
   useEffect(() => {
-    async function fetchRoads() {
-      try {
-        const response = await fetch('/api/roads');
-        if (!response.ok) {
-          throw new Error('Failed to fetch road data');
-        }
-        const data: Road[] = await response.json();
-        setRoads(data);
-      } catch (error) {
-        logError('Fetch road data', error);
-        setMapError('Could not load road data. Please try again later.');
-      } finally {
-        setIsLoading(false);
-      }
-    }
-
-    fetchRoads();
-  }, []);
-
-  // Initialize map
-  useEffect(() => {
     if (!mapContainer.current || map.current) return;
 
+    const center = userLocation || DEFAULT_COORDINATES;
+
     try {
-      // MapTiler expects [lng, lat] order for coordinates
       const [lat, lng] = center;
 
-      // Validate coordinates
       if (!isValidLatitude(lat) || !isValidLongitude(lng)) {
         logError(
           'Coordinate validation',
           new Error(ERROR_MESSAGES.INVALID_COORDINATES),
         );
-        // Use default coordinates as fallback
         const [defaultLat, defaultLng] = DEFAULT_COORDINATES;
-        const fallbackCoordinates: [number, number] = [defaultLng, defaultLat];
-        initializeMap(fallbackCoordinates);
+        initializeMap([defaultLng, defaultLat]);
         return;
       }
 
-      const validCoordinates: [number, number] = [lng, lat];
-      initializeMap(validCoordinates);
+      initializeMap([lng, lat]);
     } catch (error) {
       logError('Map initialization', error);
       setMapError(ERROR_MESSAGES.INITIALIZATION);
     }
 
     return () => {
-      try {
-        if (map.current) {
-          map.current.remove();
-          map.current = null;
-        }
-      } catch (error) {
-        logError('Cleanup', error);
+      if (map.current) {
+        map.current.remove();
+        map.current = null;
       }
     };
-  }, [center, initializeMap]);
-
-  if (isLoading) {
-    return (
-      <div className="flex items-center justify-center h-full bg-gray-100/50 dark:bg-gray-900/50">
-        <p>Loading map data...</p>
-      </div>
-    );
-  }
+  }, [userLocation, initializeMap]);
 
   if (mapError) {
     return (
-      <div className="flex items-center justify-center h-full bg-destructive/10 text-destructive p-4 rounded-md">
-        <p>{mapError}</p>
+      <div className="flex items-center justify-center h-full bg-red-100/50 dark:bg-red-900/50">
+        <p className="text-red-700">{mapError}</p>
       </div>
     );
   }
